feat: Enhance user profile management with logout functionality and settings options
