+++ app/src/main/java/com/example/aura/AuraApp.kt
@@ -9,7 +9,7 @@ class AuraApp : Application() {
 
     override fun onCreate() {
         super.onCreate()
-        container = AppContainer(this)
+        container = AppContainer(this, this)
     }
 }
