+++ app/src/main/java/com/example/aura/core/BaseViewModel.kt
@@ -3,13 +3,11 @@ package com.example.aura.core
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 
 abstract class BaseViewModel : ViewModel() {
 
     private val _isLoading = MutableStateFlow(false)
-    val isLoading: StateFlow<Boolean> get() = _isLoading
 
     protected fun <T> executeSafeCall(
         block: suspend () -> T,
