+++ app/src/main/java/com/example/aura/data/local/datasource/UserLocalDataSource.kt
@@ -7,9 +7,7 @@ import kotlinx.coroutines.flow.Flow
 class UserLocalDataSource (private val dao: UserDao) {
     suspend fun saveUser(user: UserEntity) = dao.insertUser(user)
 
-    suspend fun clearUser(){
-        dao.clearUser()
-    }
+    suspend fun clearUser() = dao.clearUser()
 
     fun getUser(): Flow<UserEntity?> = dao.getUser()
 }
