+++ app/src/main/java/com/example/aura/presentation/navigation/BottomNavBarItem.kt
@@ -3,12 +3,11 @@ package com.example.aura.presentation.navigation
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.outlined.ReceiptLong
 import androidx.compose.material.icons.outlined.AccountCircle
-import androidx.compose.material.icons.outlined.Settings
 import androidx.compose.ui.graphics.vector.ImageVector
 import com.example.aura.presentation.icons.House
+import com.example.aura.presentation.navigation.destinations.examRoute
 import com.example.aura.presentation.navigation.destinations.homeRoute
 import com.example.aura.presentation.navigation.destinations.profileRoute
-import com.example.aura.presentation.navigation.destinations.settingsRoute
 
 sealed class BottomNavBarItem(
     val label: String,
@@ -23,17 +22,11 @@ sealed class BottomNavBarItem(
     object ExamNavBarItem : BottomNavBarItem(
         label = "Exames",
         icon = Icons.AutoMirrored.Outlined.ReceiptLong,
-        route = profileRoute,
+        route = examRoute,
     )
-
     object ProfileNavBarItem : BottomNavBarItem(
         label = "Perfil",
         icon = Icons.Outlined.AccountCircle,
         route = profileRoute,
-    )
-    object SettingsNavBarItem : BottomNavBarItem(
-        label = "Config.",
-        icon = Icons.Outlined.Settings,
-        route = settingsRoute,
     )
 }
