+++ app/src/main/java/com/example/aura/presentation/navigation/destinations/ExamNavigation.kt
@@ -3,15 +3,15 @@ package com.example.aura.presentation.navigation.destinations
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
-import com.example.aura.presentation.ui.feature_profile.ProfileScreen
 import com.example.aura.di.AppContainer
+import com.example.aura.presentation.ui.feature_exam.ExamScreen
 
 const val examRoute = "examRoute"
 
 
 fun NavGraphBuilder.examScreen(navController: NavController, container: AppContainer) {
     composable(examRoute) {
-        ProfileScreen(container)
+        ExamScreen(container)
     }
 }
