--- app/src/main/java/com/example/aura/presentation/state/HomeUiState.kt
@@ -1,7 +1,0 @@
-package com.example.aura.presentation.state
-
-data class HomeUiState(
-    val isLoading: Boolean = false,
-    val message: String? = null
-)
-
