+++ app/src/main/java/com/example/aura/presentation/ui/feature_profile/ProfileScreen.kt
@@ -2,10 +2,12 @@ package com.example.aura.presentation.ui.feature_profile
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -25,7 +27,21 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.unit.dp
 import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.automirrored.filled.KeyboardArrowRight
+import androidx.compose.material.icons.automirrored.filled.Logout
+import androidx.compose.material.icons.filled.CloudUpload
+import androidx.compose.material.icons.filled.DarkMode
+import androidx.compose.material.icons.filled.Fingerprint
+import androidx.compose.material.icons.filled.Lock
+import androidx.compose.material.icons.filled.Notifications
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.DividerDefaults
 import androidx.compose.material3.ExperimentalMaterial3ExpressiveApi
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.Icon
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.layout.ContentScale
@@ -34,15 +50,21 @@ import coil.compose.rememberAsyncImagePainter
 import com.example.aura.di.AppContainer
 import com.example.aura.core.ResultWrapper
 import androidx.compose.material3.LoadingIndicator
+import androidx.compose.material3.Switch
+import kotlin.Unit
 
 @OptIn(ExperimentalMaterial3ExpressiveApi::class)
 @Composable
 fun ProfileScreen(
     container: AppContainer,
+    onLogoutSuccess: () -> Unit = {},
 ) {
-    val factory = ProfileViewModelFactory(container.userUseCases)
+    val factory = ProfileViewModelFactory(container.userUseCases, app = container.application)
     val viewModel: ProfileViewModel = viewModel(factory = factory)
-    val uiState by viewModel.uiState.collectAsState()
+    val userState by viewModel.uiUserState.collectAsState()
+
+    val darkThemeEnabled = viewModel.isDarkMode.collectAsState()
+    val cloudBackupEnabled = viewModel.isCloudBackupEnabled.collectAsState()
 
     Surface(
         modifier = Modifier.fillMaxSize(),
@@ -51,7 +73,7 @@ fun ProfileScreen(
         Box(modifier = Modifier.fillMaxSize()) {
             val scrollState = rememberScrollState()
 
-            when (uiState) {
+            when (userState) {
                 is ResultWrapper.Loading -> {
                     Box(modifier = Modifier.fillMaxWidth(), contentAlignment = Alignment.Center) {
                         LoadingIndicator(
@@ -68,7 +90,7 @@ fun ProfileScreen(
                 }
 
                 is ResultWrapper.Success -> {
-                    val user = (uiState as ResultWrapper.Success).value
+                    val user = (userState as ResultWrapper.Success).value
 
                     Column(
                         modifier = Modifier
@@ -78,16 +100,6 @@ fun ProfileScreen(
                         verticalArrangement = Arrangement.Top,
                         horizontalAlignment = Alignment.Start
                     ) {
-                        // Title centered
-                        Box(modifier = Modifier.fillMaxWidth(), contentAlignment = Alignment.Center) {
-                            Text(
-                                text = "Meu Perfil",
-                                style = MaterialTheme.typography.titleLarge,
-                                color = MaterialTheme.colorScheme.onBackground
-                            )
-                        }
-
-                        Spacer(modifier = Modifier.height(16.dp))
 
                         // Profile Card
                         Card(
@@ -156,7 +168,121 @@ fun ProfileScreen(
                         Spacer(modifier = Modifier.height(12.dp))
                         ReadOnlyField(label = "Convênio Médico", value = user?.healthInsurance ?: "")
 
+                        Spacer(modifier = Modifier.height(32.dp))
 
+                        Text(
+                            text = "Configurações",
+                            style = MaterialTheme.typography.labelLarge,
+                            color = MaterialTheme.colorScheme.onBackground
+                        )
+
+                        Spacer(modifier = Modifier.size(12.dp))
+
+                        Card(
+                            modifier = Modifier.fillMaxWidth(),
+                            shape = RoundedCornerShape(12.dp),
+                            colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.inverseOnSurface),
+                            elevation = CardDefaults.cardElevation(defaultElevation = 4.dp)
+                        ) {
+                            Column(modifier = Modifier
+                                .fillMaxWidth()
+                                .padding(12.dp)) {
+
+                                SettingRow(
+                                    icon = { Icon(Icons.Default.CloudUpload, contentDescription = null, tint = MaterialTheme.colorScheme.primary) },
+                                    title = "Backup Automático",
+                                    trailing = {
+                                        Switch(
+                                            checked = cloudBackupEnabled.value,
+                                            onCheckedChange = { viewModel.setCloudBackup(it) },
+                                        )
+                                    }
+                                )
+
+                                HorizontalDivider(
+                                    modifier = Modifier.padding(vertical = 8.dp),
+                                    thickness = DividerDefaults.Thickness,
+                                    color = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.08f)
+                                )
+
+                                SettingRow(
+                                    icon = { Icon(Icons.Default.Notifications, contentDescription = null, tint = MaterialTheme.colorScheme.primary) },
+                                    title = "Notificações",
+                                    trailing = {
+                                        Icon(Icons.AutoMirrored.Filled.KeyboardArrowRight, contentDescription = null, tint = MaterialTheme.colorScheme.onSurface)
+                                    }
+                                )
+
+                                HorizontalDivider(
+                                    modifier = Modifier.padding(vertical = 8.dp),
+                                    thickness = DividerDefaults.Thickness,
+                                    color = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.08f)
+                                )
+
+                                SettingRow(
+                                    icon = { Icon(Icons.Default.DarkMode, contentDescription = null, tint = MaterialTheme.colorScheme.primary) },
+                                    title = "Tema Escuro",
+                                    trailing = {
+                                        Switch(
+                                            checked = darkThemeEnabled.value,
+                                            onCheckedChange = { viewModel.setDarkMode(it) },
+                                        )
+                                    }
+                                )
+                            }
+                        }
+
+                        Spacer(modifier = Modifier.size(12.dp))
+
+                        Card(
+                            modifier = Modifier.fillMaxWidth(),
+                            shape = RoundedCornerShape(12.dp),
+                            colors = CardDefaults.cardColors(containerColor = MaterialTheme.colorScheme.inverseOnSurface),
+                            elevation = CardDefaults.cardElevation(defaultElevation = 4.dp)
+                        ) {
+                            Column(modifier = Modifier
+                                .fillMaxWidth()
+                                .padding(12.dp)) {
+
+                                SettingRow(
+                                    icon = { Icon(Icons.Default.Fingerprint, contentDescription = null, tint = MaterialTheme.colorScheme.primary) },
+                                    title = "Segurança e Biometria",
+                                    trailing = { Icon(Icons.AutoMirrored.Filled.KeyboardArrowRight, contentDescription = null, tint = MaterialTheme.colorScheme.onSurface) }
+                                )
+
+                                HorizontalDivider(
+                                    modifier = Modifier.padding(vertical = 8.dp),
+                                    thickness = DividerDefaults.Thickness,
+                                    color = MaterialTheme.colorScheme.onSurface.copy(alpha = 0.08f)
+                                )
+
+                                SettingRow(
+                                    icon = { Icon(Icons.Default.Lock, contentDescription = null, tint = MaterialTheme.colorScheme.primary) },
+                                    title = "Termos e Privacidade",
+                                    trailing = { Icon(Icons.AutoMirrored.Filled.KeyboardArrowRight, contentDescription = null, tint = MaterialTheme.colorScheme.onSurface) }
+                                )
+                            }
+                        }
+
+                        Spacer(modifier = Modifier.size(18.dp))
+
+                        Button(
+                            onClick = {
+                                viewModel.logout{
+                                    onLogoutSuccess()
+                                }
+                            },
+                            modifier = Modifier.fillMaxWidth(),
+                            colors = ButtonDefaults.buttonColors(
+                                containerColor = MaterialTheme.colorScheme.error.copy(alpha = 0.12f),
+                                contentColor = MaterialTheme.colorScheme.error
+                            ),
+                            shape = RoundedCornerShape(12.dp)
+                        ) {
+                            Icon(Icons.AutoMirrored.Filled.Logout, contentDescription = null)
+                            Spacer(modifier = Modifier.size(8.dp))
+                            Text(text = "Sair da Conta", style = MaterialTheme.typography.bodyLarge)
+                        }
                     }
                 }
             }
@@ -175,7 +301,6 @@ private fun ReadOnlyField(label: String, value: String) {
 
         Spacer(modifier = Modifier.height(6.dp))
 
-        // Usar Card para simular o campo read-only com as cores do tema
         Card(
             modifier = Modifier.fillMaxWidth(),
             shape = MaterialTheme.shapes.small,
@@ -188,6 +313,51 @@ private fun ReadOnlyField(label: String, value: String) {
                 color = MaterialTheme.colorScheme.onSurface,
                 modifier = Modifier.padding(horizontal = 12.dp, vertical = 12.dp)
             )
+        }
+    }
+}
+
+@Composable
+private fun SettingRow(
+    icon: @Composable () -> Unit,
+    title: String,
+    trailing: @Composable (() -> Unit)? = null
+) {
+    Row(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(vertical = 6.dp),
+        verticalAlignment = Alignment.CenterVertically,
+        horizontalArrangement = Arrangement.SpaceBetween
+    ) {
+        Row(verticalAlignment = Alignment.CenterVertically) {
+            Box(
+                modifier = Modifier
+                    .size(50.dp)
+                    .clip(CircleShape)
+                    .padding(8.dp),
+                contentAlignment = Alignment.Center
+            ) {
+                Box(modifier = Modifier
+                    .size(50.dp)
+                    .clip(CircleShape)
+                    .background(color = MaterialTheme.colorScheme.primary.copy(alpha = 0.12f))
+                )
+
+                Box(modifier = Modifier.size(35.dp), contentAlignment = Alignment.Center) {
+                    icon()
+                }
+            }
+
+            Spacer(modifier = Modifier.size(12.dp))
+
+            Text(text = title, style = MaterialTheme.typography.bodyLarge, color = MaterialTheme.colorScheme.onSurface)
+        }
+
+        if (trailing != null) {
+            Box(contentAlignment = Alignment.Center) {
+                trailing()
+            }
         }
     }
 }
