--- app/src/main/java/com/example/aura/presentation/ui/feature_settings/SettingsViewModel.kt
@@ -1,30 +1,0 @@
-package com.example.aura.presentation.ui.feature_settings
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.viewModelScope
-import com.example.aura.domain.repository.UserRepository
-import kotlinx.coroutines.launch
-
-class SettingsViewModel(
-    private val userRepository: UserRepository
-) : ViewModel(){
-    fun logout(onSuccess: () -> Unit){
-        viewModelScope.launch {
-            userRepository.logout()
-            onSuccess()
-        }
-    }
-}
-
-class SettingsViewModelFactory(
-    private val userRepository: UserRepository
-) : ViewModelProvider.Factory {
-    @Suppress("UNCHECKED_CAST")
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(SettingsViewModel::class.java)) {
-            return SettingsViewModel(userRepository) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
